+++ Messenger/Messenger/Controls/ChatControls/MessageControl.xaml.cs
@@ -1,4 +1,5 @@
-﻿using Messenger.Core.Models;
+﻿using Messenger.Commands;
+using Messenger.Core.Models;
 using Messenger.Helpers;
 using Messenger.Models;
 using Messenger.ViewModels.DataViewModels;
