+++ Messenger/Messenger/Controls/ChatControls/MessagesListControl.xaml
@@ -7,13 +7,14 @@
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:models="using:Messenger.Core.Models"
     xmlns:vm="using:Messenger.ViewModels.DataViewModels"
+    xmlns:converters="using:Messenger.Helpers.Converters"
     xmlns:helpers="using:Messenger.Helpers"
     xmlns:local="using:Messenger.Controls.ChatControls"
     mc:Ignorable="d"
     d:DesignHeight="300"
     d:DesignWidth="400">
     <UserControl.Resources>
-        <helpers:BooleanToAlignmentConverter x:Key="BooleanToAlignmentConverter" />
+        <converters:BooleanToAlignmentConverter x:Key="BooleanToAlignmentConverter" />
         <Style x:Key="OuterGrid" TargetType="Grid">
             <Setter Property="Margin" Value="0 10" />
         </Style>
