+++ Messenger/Messenger/Controls/ChatControls/SendMessageControl.xaml
@@ -7,7 +7,7 @@
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:local="using:Messenger.ViewModels"
     xmlns:icons="using:MahApps.Metro.IconPacks"
-    xmlns:converters="using:Messenger.Helpers"
+    xmlns:converters="using:Messenger.Helpers.Converters"
     mc:Ignorable="d"
     d:DesignHeight="400"
     d:DesignWidth="800">
