+++ Messenger/Messenger/Controls/Shared/UserSearchPanel.xaml
@@ -6,7 +6,7 @@
     xmlns:local="using:Messenger.Controls.Shared"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:converters="using:Messenger.Helpers"
+    xmlns:converters="using:Messenger.Helpers.Converters"
     xmlns:icons="using:MahApps.Metro.IconPacks"
     mc:Ignorable="d"
     d:DesignHeight="300"
