+++ Messenger/Messenger/Helpers/Converters/NullToVisibilityConverter.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Data;
 
-namespace Messenger.Helpers
+namespace Messenger.Helpers.Converters
 {
     /// <summary>
     /// Returns collapsed visibility option if the value is null, else visible
