--- Messenger/Messenger/Helpers/RelayCommand.cs
@@ -1,57 +1,0 @@
-﻿using System;
-using System.Windows.Input;
-
-namespace Messenger.Helpers
-{
-    public class RelayCommand : ICommand
-    {
-        private readonly Action _execute;
-
-        private readonly Func<bool> _canExecute;
-
-        public event EventHandler CanExecuteChanged;
-
-        public RelayCommand(Action execute)
-            : this(execute, null)
-        {
-        }
-
-        public RelayCommand(Action execute, Func<bool> canExecute)
-        {
-            _execute = execute ?? throw new ArgumentNullException(nameof(execute));
-            _canExecute = canExecute;
-        }
-
-        public bool CanExecute(object parameter) => _canExecute == null || _canExecute();
-
-        public void Execute(object parameter) => _execute();
-
-        public void OnCanExecuteChanged() => CanExecuteChanged?.Invoke(this, EventArgs.Empty);
-    }
-
-    public class RelayCommand<T> : ICommand
-    {
-        private readonly Action<T> _execute;
-
-        private readonly Func<T, bool> _canExecute;
-
-        public event EventHandler CanExecuteChanged;
-
-        public RelayCommand(Action<T> execute)
-            : this(execute, null)
-        {
-        }
-
-        public RelayCommand(Action<T> execute, Func<T, bool> canExecute)
-        {
-            _execute = execute ?? throw new ArgumentNullException(nameof(execute));
-            _canExecute = canExecute;
-        }
-
-        public bool CanExecute(object parameter) => _canExecute == null || _canExecute((T)parameter);
-
-        public void Execute(object parameter) => _execute((T)parameter);
-
-        public void OnCanExecuteChanged() => CanExecuteChanged?.Invoke(this, EventArgs.Empty);
-    }
-}
