+++ Messenger/Messenger/Views/MainPanels/SettingsPage.xaml
@@ -4,16 +4,16 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-    xmlns:helper="using:Messenger.Helpers" 
+    xmlns:converters="using:Messenger.Helpers.Converters" 
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
     xmlns:icons="using:MahApps.Metro.IconPacks" 
     Style="{StaticResource PageStyle}" 
     xmlns:xaml="using:Windows.UI.Xaml" 
     mc:Ignorable="d">
     <Page.Resources>
-        <helper:EnumToBooleanConverter x:Key="EnumToBooleanConverter" EnumType="ElementTheme" />
-        <helper:NullToVisibilityConverter x:Key="NullToVisConverter" />
-        <helper:ThemeToFillBrushConverter x:Key="ThemeToFillBrushConverter" />
+        <converters:EnumToBooleanConverter x:Key="EnumToBooleanConverter" EnumType="ElementTheme" />
+        <converters:NullToVisibilityConverter x:Key="NullToVisConverter" />
+        <converters:ThemeToFillBrushConverter x:Key="ThemeToFillBrushConverter" />
         <Style x:Key="LabelText" TargetType="TextBlock">
             <Setter Property="FontSize" Value="18" />
         </Style>
