+++ Messenger/Messenger/Views/NavigationPanels/ChatNavPage.xaml
@@ -11,7 +11,7 @@
     xmlns:ic="using:Microsoft.Xaml.Interactions.Core"
     xmlns:shared="using:Messenger.Controls.Shared"
     xmlns:icons="using:MahApps.Metro.IconPacks"
-    xmlns:converters="using:Messenger.Helpers"
+    xmlns:converters="using:Messenger.Helpers.Converters"
     Style="{StaticResource PageStyle}"
     mc:Ignorable="d">
     <Page.Resources>
